Fix issue 9: Possible for LightsOutTime to never be set, and handling of nil return on a variable_get leads to Deus never operating.
The issue is that the runOnce() initialization code never sets it to a
default, so on install, it isn't created. The JS UI displays "23:59" as
a default if the variable doesn't exist, but this doesn't get pushed back
into a stored value unless the user changes it. So, if the user keeps the
displayed "23:59" default, the variable is never created, and isBedtime()
takes a branch that basically says it's always bedtime, so Deus will never
be given a chance to work.

Fixed the logic, with more effort to initialize properly. Also, the old
variable LightsOutTime with its time represented in milliseconds (!) now
isn't used. Rather, a new LightsOut variable is used, which uses seconds
since midnight as its unit (easier to handle).